fix issue with pagination ins not a function; refactor code into different Components
